+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `(/ 0.0 0.0)` return NaN
 * bookmarklet always fetch latest version
 * `case-lambda` hygienic macro (code from R7RS)
+* `numerator` and `denominator` functions
 ### Bugfix
 * allow to use undefined in let
 * fix function call that return NaN